requirements.txt
nltk
textblob
newspaper3k
deep-translator
lxml
